requirements.txt
django==1.5.4
python-social-auth==0.1.14
vkontakte==1.3.2
psycopg2==2.5.1
django-jfu==1.0.7